Move SQLOrderBy[Group] extensions to own files
