+++ Sources/App/Core/Search.swift
@@ -214,48 +214,3 @@ enum Search {
         return db.raw("REFRESH MATERIALIZED VIEW \(raw: Self.schema)").run()
     }
 }
-
-
-public struct SQLOrderByGroup: SQLExpression {
-    public let orderByClauses: [SQLOrderBy]
-
-    public init(_ orderby: SQLOrderBy...) {
-        self.orderByClauses = orderby
-    }
-
-    public init(_ orderby: [SQLOrderBy]) {
-        self.orderByClauses = orderby
-    }
-
-    public func serialize(to serializer: inout SQLSerializer) {
-        guard let first = orderByClauses.first else { return }
-        first.serialize(to: &serializer)
-        for clause in orderByClauses.dropFirst() {
-            serializer.write(", ")
-            clause.serialize(to: &serializer)
-        }
-    }
-
-    func then(_ expression: SQLExpression, _ direction: SQLDirection = .ascending) -> Self {
-        SQLOrderByGroup(orderByClauses + [SQLOrderBy(expression, direction)])
-    }
-
-    func then(_ expression: SQLOrderBy) -> Self {
-        SQLOrderByGroup(orderByClauses + [expression])
-    }
-}
-
-
-extension SQLOrderBy {
-    init(_ expression: SQLExpression, _ direction: SQLDirection) {
-        self.init(expression: expression, direction: direction)
-    }
-
-    func then(_ expression: SQLExpression, _ direction: SQLDirection = .ascending) -> SQLOrderByGroup {
-        SQLOrderByGroup([self, SQLOrderBy(expression, direction)])
-    }
-
-    func then(_ expression: SQLOrderBy) -> SQLOrderByGroup {
-        SQLOrderByGroup([self, expression])
-    }
-}
